vision/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
numpy==1.26.4
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
requests==2.32.3
